Pong game separated into modules and working multiplayer
The players update correctly the position of the ball and the paddles
but there is some kind of a problem with the function that draws
the scores on the table. When the game is executed the server
stops and throws an error "Cannot call method 'emit' of undefined".
There is no emit in that function. Kinda strange???